REF: implement Manager.concat_vertical, concat_horizontal
